refactor and completed the merge
